+++ src/busOwner/services/seat-layouts.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 import { SeatLayout } from '../interfaces/seat-layout.interface';
 import { CreateSeatLayoutDto } from '../dto/create-seat-layout.dto';
 import { UpdateSeatLayoutDto } from '../dto/update-seat-layout.dto';
@@ -8,26 +8,26 @@ import { UpdateSeatLayoutDto } from '../dto/update-seat-layout.dto';
 @Injectable()
 export class SeatLayoutsService {
     constructor(@InjectModel('SeatLayout') private seatLayoutModel: Model<SeatLayout>) { }
-    
-    async create(createSeatLayoutDto: CreateSeatLayoutDto): Promise<SeatLayout> {
+
+    async create(createSeatLayoutDto: CreateSeatLayoutDto & { ownerId: Types.ObjectId }): Promise<SeatLayout> {
         console.log('Received DTO:', createSeatLayoutDto);
         const createdSeatLayout = new this.seatLayoutModel(createSeatLayoutDto);
         return createdSeatLayout.save();
     }
 
-    async findAll(): Promise<SeatLayout[]> {
-        return this.seatLayoutModel.find().exec();
+    async findAll(ownerId: Types.ObjectId): Promise<SeatLayout[]> {
+        return this.seatLayoutModel.find({ ownerId }).exec();
     }
 
-    async findOne(id: string): Promise<SeatLayout> {
-        return this.seatLayoutModel.findById(id).exec();
+    async findOne(id: string, ownerId: Types.ObjectId): Promise<SeatLayout> {
+        return this.seatLayoutModel.findById({ _id: id, ownerId }).exec();
     }
 
-    async update(id: string, updateSeatLayoutDto: UpdateSeatLayoutDto): Promise<SeatLayout> {
-        return this.seatLayoutModel.findByIdAndUpdate(id, updateSeatLayoutDto, { new: true }).exec();
+    async update(id: string, updateSeatLayoutDto: UpdateSeatLayoutDto, ownerId: Types.ObjectId): Promise<SeatLayout> {
+        return this.seatLayoutModel.findByIdAndUpdate({ _id: id, ownerId }, updateSeatLayoutDto, { new: true }).exec();
     }
 
-    async remove(id: string): Promise<SeatLayout> {
-        return this.seatLayoutModel.findByIdAndDelete(id).exec();
+    async remove(id: string, ownerId: Types.ObjectId): Promise<SeatLayout> {
+        return this.seatLayoutModel.findByIdAndDelete({ _id: id, ownerId }).exec();
     }
 }
